0.8.2

+++ src/BoxKite.Twitter/Models/Connection/TweetsContainers.cs
@@ -25,7 +25,7 @@ namespace BoxKite.Twitter.Models
         }
     }
 
-    public class DirectMessagesContainer : KeyedCollection<int, DirectMessage>
+    public class DirectMessagesContainer : KeyedCollection<long, DirectMessage>
     {
         public void SubscribeTo(IObservable<DirectMessage> stream)
         {
